requirements.txt
numpy>=1.25
pandas>=2.0
torch>=2.1
sentence-transformers>=2.2
transformers>=4.40
Pillow>=10.0
python-dotenv>=1.0
bcrypt>=4.0
streamlit>=1.24